pyproject.toml
[tool.poetry]
name = "airflow-renv-operator"
version = "1.0.0"
description = "Airflow operator for executing an R script within an renv project"
authors = ["Christopher Setzer <37311998+cmsetzer@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
apache-airflow = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
mypy = "^0.910"
pytest = "^6.2.5"

[tool.black]
line-length = 99
target-version = ["py37"]
skip-string-normalization = true

[tool.isort]
force_sort_within_sections = true
include_trailing_comma = true
line_length = 99
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
